requirements.txt
alembic==1.13.3
asgiref==3.8.1
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.68.0
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.18
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.36
starlette==0.14.2
typing_extensions==4.12.2
uvicorn==0.15.0
